refactor: add, remove하는 로직을 custom hook으로 분리
